requirements.txt
biopython==1.84
numpy==2.2.1
